pyproject.toml
[tool.poetry]
name = "learning-opengl"
version = "0.1.0"
description = ""
authors = ["Shahar Melamed <shaharonnn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyopengl = "^3.1.7"
pyopengl-accelerate = "^3.1.7"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
